requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
certifi==2022.5.18.1
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.12
    # via requests
dj-database-url==0.5.0
    # via -r requirements.in
django==4.0.5
    # via
    #   -r requirements.in
    #   django-staff-sso-client
    #   djangorestframework
django-environ==0.8.1
    # via -r requirements.in
django-rest-framework==0.1.0
    # via -r requirements.in
django-staff-sso-client==4.0.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via django-rest-framework
idna==3.3
    # via requests
oauthlib==3.2.0
    # via requests-oauthlib
psycopg2==2.9.3
    # via -r requirements.in
pytz==2022.1
    # via djangorestframework
requests==2.27.1
    # via requests-oauthlib
requests-oauthlib==1.3.1
    # via django-staff-sso-client
sentry-sdk==1.5.12
    # via -r requirements.in
sqlparse==0.4.2
    # via django
urllib3==1.26.9
    # via
    #   requests
    #   sentry-sdk
